WAITM 653: Remove unimplemented sidebar nav links (#3613)
* waitm-653: remove unimplmented sidebar links

* waitm-653: hide manual sync in avatar
